bufix/workaround: first window is never shown for me on Windows
This seems to be the only way for me to avoid the problem: explicitly
hide it and show it afterwards. On second thought this might really
be a bug in another package, as I've had the problem already years ago
on Windows with OpenGL/Lisp programs (bet never bothered too much).
Still, this is a workaround at least for cl-sdl2.